% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_aux.R
\name{.freqtab}
\alias{.freqtab}
\title{Helper function for frequency of levels by dimensions tables}
\usage{
.freqtab(
  df,
  names_eq5d = NULL,
  name_fu = NULL,
  levels_fu = NULL,
  eq5d_version = NULL,
  add_summary_problems_change = TRUE
)
}
\arguments{
\item{df}{Data frame with the EQ-5D and follow-up columns}

\item{names_eq5d}{Character vector of column names for the EQ-5D dimensions}

\item{name_fu}{Character string for the follow-up column. If NULL, no grouping is used, and the table reports for the total population.}

\item{levels_fu}{Character vector containing the order of the values in the follow-up column. 
If NULL (default value), the levels will be ordered in the order of appearance in df.}

\item{eq5d_version}{Version of the EQ-5D instrument: "3L", "5L" or
"Y3L". Matching is case-insensitive.}

\item{add_summary_problems_change}{If set to false, the resulting dataframe does not include a row on problems change.}
}
\value{
Summary data frame.
}
\description{
Helper function for frequency of levels by dimensions tables
}
