% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_aux.R
\name{.gen_colours}
\alias{.gen_colours}
\title{Generate colours for PCHC figures}
\usage{
.gen_colours(col, n)
}
\arguments{
\item{col}{A character string specifying the base colour. Only "green" or "orange" is accepted.}

\item{n}{A positive integer specifying the number of colours to generate.}
}
\value{
A vector of colours generated based on the specified base colour and number of colours.
}
\description{
This internal function generates a vector of colours based on the specified base colour. 
Currently only green and orange colours are implemented. 
The wrapper is used in Figures 2.2-2.4.
}
\keyword{internal}
