% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_aux.R
\name{.summary_table_4_4}
\alias{.summary_table_4_4}
\title{Summary wrapper for Table 4.4}
\usage{
.summary_table_4_4(df, group_by)
}
\arguments{
\item{df}{A data frame.}

\item{group_by}{A character vector of names of variables by which to group the data.}
}
\value{
A data frame with the summary statistics.
}
\description{
This internal function creates a summary of the data frame for Table 4.4. 
It groups the data by the variables specified in `group_by` and calculates various summary statistics.
}
\keyword{internal}
