% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{eq5d_profile_lfs_utility_plot}
\alias{eq5d_profile_lfs_utility_plot}
\title{eq5d_profile_lfs_utility_plot: EQ-5D values plotted against LFS}
\usage{
eq5d_profile_lfs_utility_plot(
  df,
  names_eq5d = NULL,
  name_utility = NULL,
  eq5d_version = NULL
)
}
\arguments{
\item{df}{Data frame with the EQ-5D dimension and value columns}

\item{names_eq5d}{Character vector of column names for the EQ-5D dimensions}

\item{name_utility}{Character string naming the column of pre-calculated
EQ-5D values. If \code{NULL} (default), the column \code{"utility"} is
used. The values are analysed as they stand; see
\code{\link{eq5d}} for calculating them from the dimensions.}

\item{eq5d_version}{Version of the EQ-5D instrument}
}
\value{
Summary plot and data used for plotting
}
\description{
eq5d_profile_lfs_utility_plot: EQ-5D values plotted against LFS
}
\examples{
example_data$value <- eq5d3l(example_data[, c("mo", "sc", "ua", "pd", "ad")],
                             country = "GB")
tmp <- eq5d_profile_lfs_utility_plot(
 example_data, 
 names_eq5d = c("mo", "sc", "ua", "pd", "ad"), 
 name_utility = "value",
 eq5d_version = "3L"
)
tmp$p
tmp$plot_data
}
