% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQ_functions.R
\name{make_all_EQ_indexes}
\alias{make_all_EQ_indexes}
\title{make_all_EQ_indexes}
\usage{
make_all_EQ_indexes(
  version = "5L",
  dim.names = c("mo", "sc", "ua", "pd", "ad")
)
}
\arguments{
\item{version}{Either "3L" or "5L", to signify whether 243 or 3125 states should be generated. Matching is case-insensitive.}

\item{dim.names}{A vector of dimension names to be used as names for output columns.}
}
\value{
A vector with 5-digit state indexes for all 243 (-3L) or 3125 (-5L) EQ-5D health states
}
\description{
Make a vector containing all 5-digit EQ-5D indexes for -3L or -5L version.
}
\examples{
make_all_EQ_indexes('3L')
}
