% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQ_functions.R
\name{make_all_EQ_states}
\alias{make_all_EQ_states}
\title{make_all_EQ_states}
\usage{
make_all_EQ_states(
  version = "5L",
  dim.names = c("mo", "sc", "ua", "pd", "ad"),
  append_index = FALSE
)
}
\arguments{
\item{version}{Either "3L" or "5L", to signify whether 243 or 3125 states should be generated. Matching is case-insensitive.}

\item{dim.names}{A vector of dimension names to be used as names for output columns.}

\item{append_index}{Boolean to indicate whether a column of 5-digit EQ-5D health state indexes should be added to output.}
}
\value{
A data.frame with 5 columns and 243 (-3L) or 3125 (-5L) health states
}
\description{
Make a data.frame with all health states defined by dimensions
}
\examples{
make_all_EQ_states('3L')
}
